requirements.txt
Flask==1.0.2
python-slugify==3.0.2
pytest==4.4.2
coverage==4.5.3
